git-svn-id: https://svn2.hosted-projects.com/jjinux/jenarix@237 af7818e5-df21-49d5-b9e6-ecf3e1c3d4f7

+++ trunk/src/libjx/src/vm/jx_code.c
@@ -1032,7 +1032,7 @@ jx_ob jx__code_eval_to_weak(jx_tls *tls,jx_int flags, jx_ob node, jx_ob expr)
 {
   if(flags & JX_EVAL_DEBUG_TRACE) {
     fprintf(stderr, "# trace: eval_to_weak(tls %p, flags 0x%x, ...)\n",
-            (void*)tls, flags);
+            (void*)tls, (int)flags);
   }
   if(flags & (JX_EVAL_DEBUG_DUMP_NODE)) {
     jx_jxon_dump(stderr,"#   node",node, node);
